pyproject.toml
[project]
name = "otp_cli"
description = "A RFC6238 TOTP QR Code reader and rebuilder."
readme = {file="README.rst", content-type = "text/x-rst"}
license = { text = "MIT License" }
authors = [
    {name = "Jacky Chen", email = "jacky9813@hotmail.com"}
]
dependencies = [
    "pyzbar >= 0.1.9",
    "qrcode[pil] >= 7.3.1",
    "Pillow >= 9.2.0",
    "protobuf >= 5.26.1",
    "click >= 8.1.7"
]
requires-python = ">= 3.8"

dynamic = ["version"]

[project.scripts]
# Using otp-cli as there is an application uses otp as program name within
# most of the Linux distros.
otp-cli = "totp_qrcode_analyzer.__main__:cli"

[project.urls]
Homepage = "https://github.com/jacky9813/otp-cli"
Repository = "https://github.com/jacky9813/otp-cli"
Issues = "https://github.com/jacky9813/otp-cli/issues"

[build-system]
requires = ["setuptools[toml] >= 65.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]

[tool.setuptools.package-data]
"*" = [
    "*.proto",
    "*.pyi"
]
